Sword and Skill Manager Update
Today I added a cool method that I did not knew before which is "enum". I added 4 different SwordTypes which is, Regular, Bounce, Pierce and Spin. I completed the first 3 enum and now I am planning to finish the last one. I added really cool mechanics to the Sword Throw functions and I move on to finish completely the sword functions and animations.