Black Hole Creation
Today I started to create the black hole skill and skill controller for the player. This ability will freeze the time and every enemy which is inside the black hole ability will be frost. TextMeshPro text will be displayed and hot keys of the keyboard as well. If the Player presses the hot keys correctly without any miss clicks clone will appear behind the enemies to hit them.