Update Of Skill
Today I Updated all of the skills that the player has. When the player parry enemy attacks, now, a clone of the player spawns behind the enemy and hit them. Also if the player leaves a crystal behind he can teleport to the crystal and his previous position spawns a clone and hit the enemies.