Clone Duplication + Crystal Instead of clone
Today I made a duplication function for the player's clone. When the player parry the enemy's attack, duplicates a clone and the clone attack has a small chance of creating another clone. Also I made the same function, instead of clones to spawn crystals to attack the enemies + made some small changes in the blackhole.