Health Bar Update
Today I Added A Health Bar image and placed it above the player and the enemies with a slider. I Also created the script for the Health Bar to work perfectly without any errors and have applied the damage function in it. When the Player or the Enemy takes damage the health decreases.