Update the Chill Effect
Chill Effect is now completed like ignite. When the player or the enemy receive damage the slow down their movement and animations. I am planning to fix the Shock effect and after that I will import some images for ignite, chill and Shock to be shown when the player or the enemy receives damage.