Items Update
I Updated the Items scripts. Now when the player kills an enemy he has DropChance() to drop different items. For example the enemies have 60% chance to drop leather 50% chance to drop Iron etc. Also made a prefab for the Item Object and when the enemies drop an Item it replace the image of the item.