Unique Items and Thunder Strike creation
I made A unique Item the Thunder claw and I also made a special ability to equip with which is the thunder strike. Also Applied damage function to it and when the player equips the sword and apply damage to the enemies the current enemy is damaged with lighting damage and then the Shock Strike function is applied to the nearest enemies. Last but not least, I made animations for each thunder strike and shock strike abilities.